Long-run hardening: housekeeping cadences, --soak measurement, bounded shutdown
Makes "safe to leave running for a day" a measured claim. Also brings in the pieces finished in earlier
passes that were still uncommitted: the test suite, the dashboard shell, the docs and the solution file.

capture-service
- --soak [minutes] [--soak-interval s]: samples the things that grow (CPU, working set, managed memory,
  handles, session and asset bytes, ground-truth dumps, temp-file age, index WAL, asset queue) and prints a
  verdict. It judges the steady state: slope from the second sample on, handle floor per half-window, so
  warm-up and transient waves are not misread as leaks.
- Housekeeping cadences: bounded ground-truth dump store (30-minute sweep, per-folder frame/byte caps),
  20-minute temp sweep, 3-minute WAL checkpoint + incremental_vacuum, 6-hour prune, daily budget enforcement,
  manifest hash-tracking cap, auto-vacuum conversion for existing databases.
- Session start: PrepareSessionStorage, dump-store ClearAll, budget check.
- AssetStats()/SessionBytes() are cached and refreshed off-thread, and the capture loop keeps them warm so a
  consumer that polls rarely still sees a current number.
- WriteCheckpoint drains the asset writer first: a checkpoint is the strongest claim a session makes and must
  not reference tiles that are still queued.
- Shutdown and fatal paths get a 15-second drain budget. When it expires the log flush is skipped, so the last
  interval stays atomic instead of half written.

storage-lib
- SessionLayout.SessionBytes stats each file instead of trusting directory-listing sizes: NTFS refreshes a
  file's size when it is closed, so the log being appended to reported the size it had when it was created.
- Cheaper, hardened ComputeStats and temp cleanup; monitor-entry cap; store lock.

tests, dashboard, docs
- 44 tests green. LongRunHousekeepingTests, plus pipeline tests that wait on a condition instead of a
  wall-clock window (the suite runs classes in parallel, and a short window was flaky under load) and a
  deterministic dedupe test.
- Dashboard shell: day list, scrubber, playback, jump-to-focus, settings, resource meter, tray icon, hotkey.
- docs/VALIDATION.md §7 records the long-run numbers and what building the harness found: a session size that
  was never refreshed, an unbounded shutdown drain, and Process.HandleCount reporting 2,682 handles for a
  process that an external observer saw holding 350.

+++ capture-service/Capture/CaptureEngine.Frame.cs
@@ -198,7 +198,7 @@ internal sealed partial class CaptureEngine
         _stats.LastStoreMs = Math.Round(storeMs, 2);
         _stats.LastLogMs = Math.Round(logMs, 2);
         _stats.AssetsBytesWritten = _assetWriter.BytesWritten;
-        _stats.AssetQueueDepth = _assetWriter.PendingCount;
+        _stats.AssetQueueDepth = (int)_assetWriter.PendingCount;
         return changes;
     }
 }
